memory/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helix-memory"
version = "0.1.0"
description = "Personal AI Memory Service — MCP Server + CLI"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.9.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
hmem = "helix_memory.cli.app:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["helix_memory*"]
